Add some more structure
