+++ src/server/logger.ts
@@ -1,4 +1,8 @@
+import type { NextApiRequest } from 'next'
 import winston from 'winston'
+
+export const composeDebugMessage = (req: NextApiRequest, message: string) =>
+  `Request URL: ${req.url} | Method: ${req.method} | ${message}`
 
 const logger = winston.createLogger({
   level: 'info',
